pyproject.toml
[tool.poetry]
name = "cli-llm"
version = "0.1.0"
description = ""
authors = ["OscarAdserballe <108332869+OscarAdserballe@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-generativeai = "0.5.1"
click = "8.1.3"
colored = "1.4.4"
pytesseract = "0.3.10"
tiktoken = "^0.8.0"
openai = "^1.58.1"
pyyaml = "^6.0.2"
tika = "^2.6.0"
termcolor = "^2.5.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
